minor API change: separated audio init from start
This change was necessary because in some use cases (such as the RjDj client),
you need to initialize Pd audio before loading a patch (so that samplerate~
will yield the correct result when queried via loadbang) but before starting
the audio rendering thread (so that sound files can be loaded in the
background).

I reluctantly decided to make an incompatible API change because otherwise,
the signature of the old startAudio method would have been the same as the
signature of the new initAudio method, with overlapping functionality, which
would be confusing.  I managed to confuse myself a few times before I decided
to get rid of the old startAudio methods.