fixed accelerometer coordinate system
Explanation:  Observation of RjDj patches suggests that the z-axis points
downward on iPhones.  Since I'm pretty sure that the coordinate system is
supposed to be right-handed and that the x-axis points right, I've concluded
that the way to convert between Android and iPhone accelerometer values is to
flip the sign of the y and z coordinates.