Log Bluetooth state changes. Ready to toast.
Check Bluetooth state upon initialization. If bluetooth == null log that the device doesn't support Bluetooth. If !bluetooth.isEnable log that Bluetooth is off.

Register Bluetooth receiver to listen for Bluetooth state. Log for changes and start new Bluetooth discovery if state goes to "TURNING_ON".

All of the logging places are ready to move to toast. Marked with // TODO.